% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataManagement.R
\name{checkFormat_gps}
\alias{checkFormat_gps}
\title{Check database format for GPS coordinates and bed capacities}
\usage{
checkFormat_gps(report, verbose = TRUE)
}
\arguments{
\item{report}{(list).
A list containing the base and in which will be stored reporting variables.
The database with GPS coordinates of facilities and bed capacities. The data.table should have at least the following columns:
    fID: facilityID (character)
    lat: latitude of the facility (double)
    long: longitude of the facility (double)
    beds: bed capacity in the facility (integer)}

\item{verbose}{(boolean) print diagnostic messages. Default is FALSE.}
}
\value{
Returns either an error message, or the database (modified if need be).
}
\description{
Function that performs various generic checks to ensure that the GPS database has the correct format
}
