% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FakePatientsDB.R
\name{create_fake_patientDB}
\alias{create_fake_patientDB}
\title{Create a fake patient database}
\usage{
create_fake_patientDB(n_patients = 50, n_hospital = 10,
  avg_n_stays = 3, days_since_discharge = NULL,
  length_of_stay = NULL)
}
\arguments{
\item{n_patients}{the number of different patients in the database}

\item{n_hospital}{the number of hospital present in the database}

\item{avg_n_stays}{the average number of stays per patient}

\item{days_since_discharge}{the number of days between a discharge date and an admission date (default: max(0, rnorm(1, mean = 30, sd = 10)))}

\item{length_of_stay}{the length of stay (default: max(1, rnorm(1, mean = 5, sd = 3) )}
}
\value{
a data.table containing all patients stays
}
\description{
Create a fake patient database
}
