% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FakePatientsDB.R
\name{create_fake_subjectDB_clustered}
\alias{create_fake_subjectDB_clustered}
\title{Create a fake subject database with clustering}
\usage{
create_fake_subjectDB_clustered(
  n_subjects = 50,
  n_facilities = 10,
  avg_n_stays = 3,
  days_since_discharge = NULL,
  length_of_stay = NULL,
  n_clusters = 3
)
}
\arguments{
\item{n_subjects}{the number of different subjects in the database}

\item{n_facilities}{the number of facility present in the database}

\item{avg_n_stays}{the average number of stays per subject}

\item{days_since_discharge}{the number of days between a discharge date and an admission date (default: max(0, rnorm(1, mean = 30, sd = 10)))}

\item{length_of_stay}{the length of stay (default: max(1, rnorm(1, mean = 5, sd = 3) )}

\item{n_clusters}{the number of cluster in the network}
}
\value{
a data.table containing all subjects stays
}
\description{
Create a fake subject database with clustering
}
