% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataManagement.R
\name{deleteErrorRecords}
\alias{deleteErrorRecords}
\title{Function that removes error records, either just one record or entire patient.}
\usage{
deleteErrorRecords(base, theErrors, deleteErrors = NULL,
  patientID = "pID", hospitalID = "hID", disDate = "Ddate",
  admDate = "Adate", verbose = FALSE, ...)
}
\arguments{
\item{base}{(data.table).
A patient discharge database, in the form of a data.table. The data.table should have at least the following columns:
    pID: patientID (character)
    hID: hospitalID (character)
    Adate: admission date (POSIXct, but character can be converted to POSIXct)
    Ddate: discharge date (POSIXct, but character can be converted to POSIXct)}

\item{theErrors}{a list of indeces for the records to be deleted in the base datatable}

\item{deleteErrors}{(character) How incorrect records should be deleted: 
"record" deletes just the incorrect record
"patient" deletes all records of each patient with one or more incorrect records.}

\item{patientID}{(charachter) the columns name containing the patient ID. Default is "pID"}

\item{hospitalID}{(charachter) the columns name containing the hospital ID. Default is "hID"}

\item{disDate}{(charachter) the columns name containing the discharge date. Default is "Ddate"}

\item{admDate}{(charachter) the columns name containing the admission date. Default is "Adate"}

\item{verbose}{(boolean) print diagnostic messages. Default is FALSE.}

\item{...}{other parameters passed on to internal functions}
}
\value{
The adjusted database as a data.table
}
\description{
Function that removes error records, either just one record or entire patient.
}
