% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NetworkAnalysis.R
\name{get_betweenness}
\alias{get_betweenness}
\title{Compute the betweeness centrality}
\usage{
get_betweenness(graph)
}
\arguments{
\item{graph}{an igraph object}
}
\value{
a data.table containing the centrality measure
}
\description{
Compute the betweeness centrality
}
