% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NetworkAnalysis.R
\name{get_closeness}
\alias{get_closeness}
\title{Compute closeness}
\usage{
get_closeness(graph, modes = "total")
}
\arguments{
\item{graph}{an igraph object}

\item{modes}{option passed on to igraph::closeness : "out", "in", "all", "total"}
}
\value{
a data.table containing the closeness measure
}
\description{
Compute one or several closeness measure for hospital networks.
}
\seealso{
\code{\link[igraph]{closeness}}
}
