% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NetworkAnalysis.R
\name{get_degree}
\alias{get_degree}
\title{Compute the degree of each nodes in the network}
\usage{
get_degree(graph, modes = c("in", "out", "total"))
}
\arguments{
\item{graph}{an igraph object}

\item{modes}{the type of degree: "in", "out", "total"}
}
\value{
a data.table of nodes degree
}
\description{
Compute the degree of each nodes in the network
}
