% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NetworkAnalysis.R
\name{get_eigenvector_centrality}
\alias{get_eigenvector_centrality}
\title{Compute the eigenvector centrality}
\usage{
get_eigenvector_centrality(graph)
}
\arguments{
\item{graph}{An igraph object.}
}
\value{
A data.table containing eigenvector centrality for each node.
}
\description{
Compute the eigenvector centrality
}
\seealso{
\code{\link[igraph]{eigen_centrality}}
}
