% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NetworkAnalysis.R
\name{get_hubs_bycluster}
\alias{get_hubs_bycluster}
\title{Function computing hub scores of nodes by group}
\usage{
get_hubs_bycluster(graphs, name, ...)
}
\arguments{
\item{graphs}{A list of igraph graphs, one for each group within which the hub scores will be computed}

\item{name}{[character (1)] The name of grouping variable (used only for naming the column of the DT)}

\item{...}{Optional arguments to be passed to igraph function 'hub_score()'}
}
\description{
Function computing hub scores of nodes by group
}
