% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NetworkAnalysis.R
\name{get_hubs_global}
\alias{get_hubs_global}
\title{Function computing hub scores for each node. If bycluster = TRUE, hub scores are computed by cluster}
\usage{
get_hubs_global(graph, ...)
}
\arguments{
\item{graph}{An igraph graph}
}
\description{
Function computing hub scores for each node. If bycluster = TRUE, hub scores are computed by cluster
}
