% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NetworkAnalysis.R
\name{get_matrix_bycluster}
\alias{get_matrix_bycluster}
\title{Function returning matrices of transfers within each by clusters}
\usage{
get_matrix_bycluster(mat, DT, clusters)
}
\arguments{
\item{mat}{The adjacency matrix of the network}

\item{DT}{A data table with at least a column 'node' and a factor column identifying the node's cluster}

\item{clusters}{A unique character vector of the name of the column identifying the nodes' clusters}
}
\description{
Function returning matrices of transfers within each by clusters
}
