% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NetworkAnalysis.R
\name{get_reciprocity}
\alias{get_reciprocity}
\title{Compute reciprocity for a directed network}
\usage{
get_reciprocity(graph)
}
\arguments{
\item{graph}{an igraph object (directed graph)}
}
\value{
a data.table containing reciprocity values for each node
}
\description{
Compute reciprocity for a directed network
}
\seealso{
\code{\link[igraph]{reciprocity}}
}
