% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NetworkAnalysis.R
\name{getMetrics}
\alias{getMetrics}
\title{Function computing different network analysis metrics}
\usage{
getMetrics(network, mode = "directed", weighted = TRUE,
  transfers = TRUE, metrics = c("degree", "closeness", "betweenness"),
  clusters = c("cluster_fast_greedy", "cluster_infomap"),
  hubs = "all_clusters", options = list(degree = list(modes = c("in",
  "out", "total")), closeness = list(modes = "total"), betweenness =
  list(), cluster_fast_greedy = list(undirected = "collapse"),
  cluster_infomap = list(undirected = "collapse")))
}
\arguments{
\item{network}{The network to analyse. Must be a square adjacency matrix (n*n).}

\item{metrics}{A list of the metrics to compute}

\item{options}{A named list of options to be passed to the igraph functions}
}
\description{
Function computing different network analysis metrics
}
