% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HospiNet.R
\name{HospiNet}
\alias{HospiNet}
\title{Constructor for the HospiNet object}
\usage{
HospiNet(matrix, edgelist, window_threshold, nmoves_threshold, noloops)
}
\arguments{
\item{matrix}{}

\item{edgelist}{}

\item{window_threshold}{}

\item{nmoves_threshold}{}

\item{noloops}{}
}
\description{
Constructor for the HospiNet object
}
