% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NetworkBuilding.R
\name{hospinet_from_patient_database}
\alias{hospinet_from_patient_database}
\title{Create HospiNet object from patient database}
\usage{
hospinet_from_patient_database(base, patientID = "pID",
  hospitalID = "hID", admDate = "Adate", disDate = "Ddate",
  noloops = TRUE, window_threshold = 365, nmoves_threshold = NULL,
  verbose = FALSE)
}
\arguments{
\item{base}{}

\item{patientID}{}

\item{hospitalID}{}

\item{admDate}{}

\item{disDate}{}

\item{noloops}{}

\item{window_threshold}{}

\item{nmoves_threshold}{}
}
\value{
a HospiNet object
}
\description{
Create HospiNet object from patient database
}
\examples{
mydb = create_fake_patientDB(n_patients = 100, n_hospital = 5)
mat = hospinet_from_patient_database(base = mydb)
mat

}
