% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NetworkBuilding.R
\name{hospinet_from_subject_database}
\alias{hospinet_from_subject_database}
\title{Create HospiNet object from subject database}
\usage{
hospinet_from_subject_database(base, window_threshold = 365,
  count_option = "successive", condition = "dates", noloops = TRUE,
  nmoves_threshold = NULL, flag_vars = NULL, flag_values = NULL,
  subjectID = "sID", facilityID = "fID", admDate = "Adate",
  disDate = "Ddate", create_MetricsTable = TRUE, verbose = FALSE)
}
\arguments{
\item{base}{(hospinet.base) A database of records of stays of subjects in
facilities. This can be obtained using the function \code{\link{checkBase}}.}

\item{window_threshold}{(numeric)
A threshold for the number of days between discharge and admission to be counted as a transfer. Set to 0 for same day transfer, default is 365 days.}

\item{noloops}{(boolean).
Should transfers within the same nodes (loops) be kept or set to 0. Defaults to TRUE, removing loops (setting matrix diagonal to 0).}

\item{nmoves_threshold}{(numeric)
A threshold for the minimum number of subject transfer between two facilities. Set to NULL to deactivate, default to NULL.}

\item{subjectID, facilityID, admDate, disDate}{(character)
Change the default names of the base columns.}

\item{create_MetricsTable}{(boolean)
Should the metrics table be created along with the network. Setting to FALSE will speed up the results. Default is TRUE.}

\item{verbose}{TRUE to print computation steps}
}
\value{
The function returns a HospiNet object.
}
\description{
This function creates a HospiNet object from the database containing subjects stays.
}
\examples{
#TODO

}
\seealso{
\code{\link{HospiNet}}
}
