% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HospitalNetwork-package.R
\docType{package}
\name{HospitalNetwork-package}
\alias{HospitalNetwork}
\alias{HospitalNetwork-package}
\title{HospitalNetwork: Building Networks of Hospitals Through Patients Transfers}
\description{
Set of tools to help interested researchers to build hospital networks from data on hospitalized patients transferred between hospitals. Methods provided have been used in Donker T, Wallinga J, Grundmann H. (2010) \doi{10.1371/journal.pcbi.1000715}, and Nekkab N, Crépey P, Astagneau P, Opatowski L, Temime L. (2020) \doi{10.1038/s41598-020-71212-6}.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://pascalcrepey.github.io/HospitalNetwork/}
  \item Report bugs at \url{https://github.com/PascalCrepey/HospitalNetwork/issues}
}

}
\author{
\strong{Maintainer}: Pascal Crépey \email{pascal.crepey@ehesp.fr} [copyright holder]

Authors:
\itemize{
  \item Tjibbe Donker \email{tjibbe.donker@uniklinik-freiburg.de}
  \item Clément Massonnaud \email{clement.massonnaud@gmail.com}
  \item Michael Lydeamore \email{michael.lydeamore@monash.edu}
}

}
\keyword{internal}
