% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_construct_network.R
\name{mod_construct_network_ui}
\alias{mod_construct_network_ui}
\alias{mod_construct_network_server}
\title{mod_construct_network_ui and mod_construct_network_server}
\usage{
mod_construct_network_ui(id)

mod_construct_network_server(input, output, session, base)
}
\arguments{
\item{id}{shiny id}

\item{input}{internal}

\item{output}{internal}

\item{session}{internal}
}
\description{
A shiny Module.
}
\keyword{internal}
