% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_database.R
\name{mod_database_ui}
\alias{mod_database_ui}
\alias{mod_database_server}
\title{mod_database_ui and mod_database_server}
\usage{
mod_database_ui(id)

mod_database_server(input, output, session, base)
}
\arguments{
\item{id}{shiny id}

\item{input}{internal}

\item{output}{internal}

\item{session}{internal}
}
\description{
A shiny Module.
}
\keyword{internal}
