% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_loadNcheck.R
\name{mod_loadNcheck_ui}
\alias{mod_loadNcheck_ui}
\alias{mod_loadNcheck_server}
\title{mod_loadNcheck_ui and mod_loadNcheck_server}
\usage{
mod_loadNcheck_ui(id)

mod_loadNcheck_server(input, output, session, parent)
}
\arguments{
\item{id}{shiny id}

\item{input}{internal}

\item{output}{internal}

\item{session}{internal}
}
\description{
Module to load the database, perform checks, and print messages, warnings, and errors. Server returns the cleaned database.
}
\keyword{internal}
