% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_map.R
\name{mod_map_ui}
\alias{mod_map_ui}
\alias{mod_map_server}
\title{map UI Function}
\usage{
mod_map_ui(id)

mod_map_server(input, output, session, net)
}
\arguments{
\item{id}{shiny id}

\item{input}{internal}

\item{output}{internal}

\item{session}{internal}
}
\description{
A shiny Module to create the map of the network using leaflet.
}
\keyword{internal}
