% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_network.R
\name{mod_network_ui}
\alias{mod_network_ui}
\alias{mod_network_server}
\title{mod_network_ui and mod_network_server}
\usage{
mod_network_ui(id)

mod_network_server(input, output, session, net)
}
\arguments{
\item{id}{shiny id}

\item{input}{internal}

\item{output}{internal}

\item{session}{internal}
}
\description{
A shiny Module.
}
\keyword{internal}
