% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_statmet.R
\name{mod_statmet_ui}
\alias{mod_statmet_ui}
\alias{mod_statmet_server}
\title{mod_statmet_ui and mod_statmet_server}
\usage{
mod_statmet_ui(id)

mod_statmet_server(input, output, session, net)
}
\arguments{
\item{id}{shiny id}

\item{input}{internal}

\item{output}{internal}

\item{session}{internal}
}
\description{
A shiny Module.
}
\keyword{internal}
