% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataSummary.R
\name{per_facility_summary}
\alias{per_facility_summary}
\title{Function that extracts summary statistics from entire database}
\usage{
per_facility_summary(base, subjectID = "sID", facilityID = "fID",
  disDate = "Ddate", admDate = "Adate", verbose = FALSE, ...)
}
\arguments{
\item{base}{(data.table).
A subject discharge database, in the form of a data.table. The data.table should have at least the following columns:
    sID: subjectID (character)
    fID: facilityID (character)
    Adate: admission date (date)
    Ddate: discharge date (date)}

\item{subjectID}{(charachter) the columns name containing the subject ID. Default is "sID"}

\item{facilityID}{(charachter) the columns name containing the facility ID. Default is "fID"}

\item{disDate}{(charachter) the columns name containing the discharge date. Default is "Ddate"}

\item{admDate}{(charachter) the columns name containing the admission date. Default is "Adate"}

\item{verbose}{(boolean) print diagnostic messages. Default is TRUE.}

\item{...}{other parameters passed on to internal functions}
}
\value{
a data table with one row per facility, showing mean LOS, number of subjects, and number of admissions
}
\description{
Function that extracts summary statistics from entire database
}
