% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataManagement.R
\name{adjust_overlapping_stays}
\alias{adjust_overlapping_stays}
\title{Check and fix overlapping admissions.}
\usage{
adjust_overlapping_stays(report, subjectID = "sID", facilityID = "fID",
  admDate = "Adate", disDate = "Ddate", maxIteration = 25,
  verbose = FALSE, retainAuxData = TRUE, ...)
}
\arguments{
\item{subjectID}{(charachter) the columns name containing the subject ID. Default is "sID".}

\item{facilityID}{(charachter) the columns name containing the facility ID. Default is "fID".}

\item{admDate}{(charachter) the columns name containing the admission date. Default is "Adate".}

\item{disDate}{(charachter) the columns name containing the discharge date. Default is "Ddate".}

\item{maxIteration}{(integer) the maximum number of times the function will try and remove overlapping admissions.}

\item{verbose}{(boolean) print diagnostic messages. Default is FALSE.}

\item{...}{other parameters passed on to internal functions}

\item{base}{(data.table).
A patient discharge database, in the form of a data.table. The data.table should have at least the following columns:
    sID: subjectID (character)
    fID: facilityID (character)
    Adate: admission date (POSIXct, but character can be converted to POSIXct)
    Ddate: discharge date (POSIXct, but character can be converted to POSIXct)}
}
\value{
The corrected database as data.table.
}
\description{
This function checks if a discharge (n) is not later than the next (n+1) admission.
If this is the case, it sets the date of discharge n to date of discharge n+1, and creates an extra record running from discharge n+1 to discharge n.
If the length of stay of this record is negative, it removes it.
It is possible that one pass of this algorithm doesn't clear all overlapping admissions (e.g. when one admission overlaps with more than one other admission), it is therefore iterated until no overlapping admissions are found.
Returns the corrected database.
}
