% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataSummary.R
\name{all_admissions_summary}
\alias{all_admissions_summary}
\title{Summary statistics on entire database}
\usage{
all_admissions_summary(base, patientID = "pID", hospitalID = "hID",
  disDate = "Ddate", admDate = "Adate", verbose = FALSE, ...)
}
\arguments{
\item{base}{(data.table).
A patient discharge database, in the form of a data.table. The data.table should have at least the following columns:
    pID: patientID (character)
    hID: hospitalID (character)
    Adate: admission date (date)
    Ddate: discharge date (date)}

\item{patientID}{(charachter) the columns name containing the patient ID. Default is "pID"}

\item{hospitalID}{(charachter) the columns name containing the hospital ID. Default is "hID"}

\item{disDate}{(charachter) the columns name containing the discharge date. Default is "Ddate"}

\item{admDate}{(charachter) the columns name containing the admission date. Default is "Adate"}

\item{verbose}{(boolean) print diagnostic messages. Default is TRUE.}

\item{...}{other parameters passed on to internal functions}
}
\value{
a list of summary statistics:
        -   meanLOS: The mean length of stay, in days
        -   meanTBA: The mean time between admissions, in days
        -   totalAdmissions: Total number of admissions (i.e. number of records in the database)
        -   numPatients: Number of unique patients
        -   numHospitals: Number of unique hospitals
        -   LOSdistribution: Distribution of length of stay
        -   TBAdistribution: Distribution of time between admissions
}
\description{
Function that extracts summary statistics from entire database
}
