% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataManagement.R
\name{checkBase}
\alias{checkBase}
\title{General check function}
\usage{
checkBase(base, deleteMissing = "no", convertDates = FALSE,
  dateFormat = "", bestGuess = FALSE, overnight = FALSE,
  deleteErrors = "", patientID = "pID", hospitalID = "hID",
  disDate = "Ddate", admDate = "Adate", ...)
}
\arguments{
\item{base}{(data.table).
A patient discharge database, in the form of a data.table. The data.table should have at least the following columns:
    pID: patientID (character)
    hID: hospitalID (character)
    Adate: admission date (POSIXct, but character can be converted to POSIXct)
    Ddate: discharge date (POSIXct, but character can be converted to POSIXct)}

\item{deleteMissing}{(character) How to handle records that contain a missing value in at least one of the four mandatory variables:
"no" (default): do not delete. Stops the function with an error message.
"record": deletes just the incorrect record.
"patient": deletes all records of each patient with one or more incorrect records.}

\item{deleteErrors}{(character) How incorrect records should be deleted: 
"record" deletes just the incorrect record
"patient" deletes all records of each patient with one or more incorrect records.}
}
\value{
The adjusted database as a data.table
}
\description{
Function that performs various checks to ensure the database is correctly formatted, and adjusts overlapping patient records
}
