% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataManagement.R
\name{checkDates}
\alias{checkDates}
\title{Check and fix the dates in patient database}
\usage{
checkDates(base, patientID = "pID", hospitalID = "hID",
  admDate = "Adate", disDate = "Ddate", convertDates = FALSE,
  dateFormat = NULL, deleteErrors = NULL, verbose = FALSE, ...)
}
\arguments{
\item{base}{(data.table).
A patient discharge database, in the form of a data.table. The data.table should have at least the following columns:
    pID: patientID (character)
    hID: hospitalID (character)
    Adate: admission date (POSIXct, but character can be converted to POSIXct)
    Ddate: discharge date (POSIXct, but character can be converted to POSIXct)}

\item{convertDates}{(boolean) indicating if dates need to be converted to POSIXct if they are not}

\item{dateFormat}{(character) giving the input format of the date character string}

\item{deleteErrors}{(character) indicating the way incorrect records should be deleted: 
"record" deletes just the incorrect record
"patient" deletes all records of each patient with one or more incorrect records.}

\item{verbose}{(boolean) print diagnostic messages. Default is FALSE.}

\item{...}{other parameters passed on to internal functions}
}
\value{
The corrected database as data.table.
}
\description{
Function that checks the dates in the data base and excludes incorrect records, returns the corrected database
}
