% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataManagement.R
\name{checkFormat}
\alias{checkFormat}
\title{Check database format}
\usage{
checkFormat(base, deleteMissing = "no")
}
\arguments{
\item{base}{(data.table)
A patient discharge database, in the form of a data.table. The data.table should have at least the following columns:
    pID: patientID (character)
    hID: hospitalID (character)
    Adate: admission date (POSIXct, but character can be converted to POSIXct)
    Ddate: discharge date (POSIXct, but character can be converted to POSIXct)}

\item{deleteMissing}{(character) How to handle records that contain a missing value in at least one of the four mandatory variables:
"no" (default): do not delete. Stops the function with an error message.
"record": deletes just the incorrect record.
"patient": deletes all records of each patient with one or more incorrect records.}
}
\value{
Returns either an error message, or the database (modified if need be).
}
\description{
Function that performs various generic checks to ensure that the database has the correct format
}
