% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NetworkBuilding.R
\name{edgelist_from_patient_database}
\alias{edgelist_from_patient_database}
\title{createMatrix
This function creates the adjacency matrix of the network from a patient discharge database. Can also return an edge list instead of the adjacency matrix.}
\usage{
edgelist_from_patient_database(base, patientID = "ID",
  hospitalID = "FINESS", noloops = TRUE, threshold = NULL)
}
\arguments{
\item{base}{(data.table).
A patient discharge database, in the form of a data.table. The data.table should have at least the following columns:
    patientID (character)
    stayOrder (character) OR dateBeginStay (date) AND dateEndStay (date)
    hospitalID (character)
    mode_entree (character, one of "mutation", "transfert" or "domicile")
    mode_sortie (character, one of "mutation", "transfert", "domicile" or "deces")}

\item{noloops}{(boolean).
Should transfers within the same nodes (loops) be kept or set to 0. Defaults to TRUE, removing loops (setting matrix diagonal to 0).}

\item{threshold}{(numeric)
A threshold below which the node should be removed from the network. The threshold correspond to the sum of in AND out transfers for the node. Defaults to NULL.}
}
\value{
The adjency matrix or the edge list in the form of a data.table.
}
\description{
createMatrix
This function creates the adjacency matrix of the network from a patient discharge database. Can also return an edge list instead of the adjacency matrix.
}
