% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NetworkAnalysis.R
\name{get_reciprocity}
\alias{get_reciprocity}
\title{Compute the reciprocity}
\usage{
get_reciprocity(graph)
}
\arguments{
\item{graph}{An igraph object (directed graph).}
}
\value{
A data.table containing reciprocity for each node.
}
\description{
Compute the reciprocity
}
\seealso{
\code{\link[igraph]{reciprocity}}
}
