% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HospiNet.R
\docType{data}
\name{HospiNet}
\alias{HospiNet}
\title{Constructor for the HospiNet object}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
HospiNet
}
\arguments{
\item{edgelist}{(data.table) the list of edges (origin, target) and their associated number of movements (N)}

\item{window_threshold}{the window threshold used to compute the network}

\item{nmoves_threshold}{the nmoves threshold used to compute the network}

\item{noloops}{TRUE if loops have been removed}
}
\value{
The constructor returns a HospiNet S3 object.
}
\description{
Constructor for the HospiNet object
}
\keyword{datasets}
