% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HospiNet.R
\docType{class}
\name{HospiNet}
\alias{HospiNet}
\title{Class providing the HospiNet object with its methods}
\format{\code{\link{R6Class}} object.}
\usage{
HospiNet
}
\value{
Object of \code{\link{R6Class}} with methods for accessing hospital networks.
}
\description{
Class providing the HospiNet object with its methods
}
\section{Fields}{

\describe{
\item{\code{edgelist}}{(data.table) the list of edges (origin, target) and their associated number of movements (N) (read-only)}

\item{\code{matrix}}{(matrix) the transfer matrix (active binding, read-only)}

\item{\code{igraph}}{(igraph) the igraph object corresponding to the network (active binding, read-only)}

\item{\code{n_hospitals}}{the number of hospitals in the network (read-only)}

\item{\code{n_movements}}{the total number of patient movements in the network (read-only)}

\item{\code{window_threshold}}{the window threshold used to compute the network (read-only)}

\item{\code{nmoves_threshold}}{the nmoves threshold used to compute the network (read-only)}

\item{\code{noloops}}{TRUE if loops have been removed (read-only)}

\item{\code{degrees}}{number of connections for each hospitals (total, in, and out)}

\item{\code{hist_degrees}}{histogram data of the number of connections per hospital}
}}

\section{Methods}{

\describe{
  \item{\code{new(edgelist, 
window_threshold,
nmoves_threshold,
noloops)}}{This method is used to create an object of this class with \code{edgelist} as the necessary information to create the network.
The other arguments \code{window_threshold}, \code{nmoves_threshold}, and \code{noloops} are specific to the \code{edgelist} and need to be provided.
For ease of use, it is preferable to use the function \code{\link{hospinet_from_patient_database}}}
  \item{\code{print()}}{This method prints basic information about the object.}
  \item{\code{plot(type = "matrix")}}{This method plots the network matrix by default. 
  The argument \code{type} can take the following values: 
  \describe{
  \item{matrix}{print the network matrix,}
  \item{clustered_matrix}{identify and print cluster(s) in the matrix using the infomap algorithm (from igraph).}
  }
  }
}
}

\examples{
mydbsmall = create_fake_patientDB(n_patients = 1000, n_hospital = 10)
  
hn = hospinet_from_patient_database(base = mydbsmall)

hn

plot(hn)
plot(hn, type = "clustered_matrix")

}
\keyword{data}
