% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NetworkBuilding.R
\name{matrix_from_edgelist}
\alias{matrix_from_edgelist}
\title{Function that creates an adjacency matrix from an edgelist}
\usage{
matrix_from_edgelist(edgelist, origin = "FINESS_DEF_1",
  target = "FINESS_DEF_2", value.var = "NB_transferts",
  format.long = F)
}
\arguments{
\item{origin}{the name of the variable containing the origin hospital id}

\item{target}{the name of the variable containing the target hospital id}

\item{value.var}{the name of the variable containing the number of transfers}
}
\description{
Function that creates an adjacency matrix from an edgelist
}
